# docs/openapi.yaml
openapi: 3.0.3
info:
  title: TwelveGo API
  description: "TwelveGo 後端 API 文件，支援 Google OAuth、檔案上傳、會員功能、購物車等。"
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: "本地開發環境"
paths:
...
  /auth/login:
    post:
      summary: 用戶登入
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                google_id:
                  type: string
                  example: "google_abcdef"
      responses:
        "200":
          description: 登入成功
        "401":
          description: 認證失敗
  /auth/logout:
    post:
      summary: 用戶登出
      operationId: logout
      responses:
        "200":
          description: 登出成功
        "400":
          description: 無效的會話
...